+++ src/templating.ts
@@ -1,6 +1,6 @@
 import { format as strFormat } from "util";
 
-import { CommitData } from "./types";
+import { OptimizeFileData, OptimizeProjectData } from "./types";
 
 import { getFileSizeInKB } from "./utils/file-size";
 import { toPercentage } from "./utils/percentages";
@@ -8,18 +8,15 @@ import { toPercentage } from "./utils/percentages";
 
 const FILES_LIST_EXP = /\{\{\s*filesList\s*\}\}/;
 const FILES_TABLE_EXP = /\{\{\s*filesTable\s*\}\}/;
-const IGNORED_COUNT_EXP = /\{\{\s*ignoredCount\s*\}\}/;
 const OPTIMIZED_COUNT_EXP = /\{\{\s*optimizedCount\s*\}\}/;
 const SKIPPED_COUNT_EXP = /\{\{\s*skippedCount\s*\}\}/;
 const SVG_COUNT_EXP = /\{\{\s*svgCount\s*\}\}/;
 const WARNINGS_EXP = /\{\{\s*warnings\s*\}\}/;
 
-const FILE_DATA_KEY = "fileData";
-const IGNORED_COUNT_KEY = "ignoredCount";
+const FILE_DATA_KEY = "files";
 const OPTIMIZED_COUNT_KEY = "optimizedCount";
 const SKIPPED_COUNT_KEY = "skippedCount";
 const SVG_COUNT_KEY = "svgCount";
-const WARNINGS_KEY = "warnings";
 
 const FILES_TABLE_HEADER =
   "| Filename | Before | After | Improvement |\n" +
@@ -29,33 +26,25 @@ const FILES_TABLE_ROW = "| %s | %s KB | %s KB | %s%% |\n";
 const formatters = [
   {
     key: FILE_DATA_KEY,
-    fn: (template: string, value: CommitData["fileData"]): string => {
-      const paths: string[] = value.optimized.map((fileData) => fileData.path);
+    fn: (template: string, values: OptimizeFileData[]): string => {
+      const paths: string[] = values.map((fileData) => fileData.path);
       return template.replace(FILES_LIST_EXP, "- " + paths.join("\n- "));
     },
   },
   {
     key: FILE_DATA_KEY,
-    fn: (template: string, value: CommitData["fileData"]): string => {
+    fn: (template: string, values: OptimizeFileData[]): string => {
       let table = FILES_TABLE_HEADER;
-      for (const optimizedSvg of value.optimized) {
-        const originalSvg = value.original.find((originalSvg) => {
-          return originalSvg.path === optimizedSvg.path;
-        });
-
-        if (originalSvg === undefined) {
-          throw new Error("Original version of optimized SVG missing");
-        }
-
-        const originalSize: number = getFileSizeInKB(originalSvg.content);
-        const optimizedSize: number = getFileSizeInKB(optimizedSvg.content);
+      for (const value of values) {
+        const originalSize: number = getFileSizeInKB(value.contentBefore);
+        const optimizedSize: number = getFileSizeInKB(value.contentAfter);
 
         const reduced: number = (originalSize - optimizedSize) / originalSize;
         const reducedPercentage: number = -1 * toPercentage(reduced);
 
         table += strFormat(
           FILES_TABLE_ROW,
-          optimizedSvg.path,
+          value.path,
           originalSize,
           optimizedSize,
           reducedPercentage,
@@ -63,12 +52,6 @@ const formatters = [
       }
 
       return template.replace(FILES_TABLE_EXP, table);
-    },
-  },
-  {
-    key: IGNORED_COUNT_KEY,
-    fn: (template: string, value: number): string => {
-      return template.replace(IGNORED_COUNT_EXP, value.toString());
     },
   },
   {
@@ -89,38 +72,27 @@ const formatters = [
       return template.replace(SVG_COUNT_EXP, value.toString());
     },
   },
-  {
-    key: WARNINGS_KEY,
-    fn: (template: string, value: string[]): string => {
-      let warnings = "";
-      if (value.length > 0) {
-        warnings = value.reduce((acc, cur) => `${acc}\n- ${cur}`, "WARNINGS:");
-      }
-
-      return template.replace(WARNINGS_EXP, warnings);
-    },
-  },
 ];
 
-function formatAll(
-  template: string,
-  data: CommitData,
-  exclude: string[] = [],
-): string {
-  for (const { key, fn } of formatters) {
-    if (!exclude.includes(key)) {
-      // eslint-disable-next-line security/detect-object-injection
-      template = fn(template, data[key]);
-    }
+function injectWarnings(template: string, values: string[]): string {
+  let warnings = "";
+  if (values.length > 0) {
+    warnings = values.reduce((acc, cur) => `${acc}\n- ${cur}`, "WARNINGS:");
   }
 
-  return template;
+  return template.replace(WARNINGS_EXP, warnings);
 }
 
 
 export function formatComment(
-  commentTemplate: string,
-  data: CommitData,
+  template: string,
+  data: OptimizeProjectData,
+  warnings: string[],
 ): string {
-  return formatAll(commentTemplate, data);
+  for (const { key, fn } of formatters) {
+    // eslint-disable-next-line security/detect-object-injection
+    template = fn(template, data[key]);
+  }
+
+  return injectWarnings(template, warnings);
 }